Movie and Weather GUIs
Weather GUI is completed, able to tell the temperature, windspeed, location and name of any city in the US by typing it into the appropriate bars.

MovieAPI is very nearly completed. Able to show the poster, title, date and popularity score, along with the general overview. The only problem is that the scroll bar is not currently working